More Redux
Connected all the moving parts of redux for the products in the store

Reducers - change and modify state in a predictable way based on action type. The reducer function recieves the action.

Actions - The action object is dispatched to the reducer function. The action has a type that tells the reducer how to modify the state and the payload that holds data the reducer can use.

Dispatchers - Dispatch passes the action to the reducer function

We keep our actions, reducers, and constants in separate folders for organization. They are all connected in the redux store where all the reducers are combined with the combineReducers function from 'redux' .

The global redux state and reducers can be accessed using the useSelector and useDispatch hooks from 'react-redux'