Product ID bugfix + Order Reducer
Created redux state for creating orders so when order is placed the order is sent to the server where it validates the order through the order model and saves the order to the database. The server either responds with an error or it sends a json response o the order document it created. The user is then redirected to the '/orders/:id' which has not been implemented yet.

Product ID bugfix, productID was incorrectly labeled preventing the cart items from being read by the database when it expected to see just 'product'