POST /api/products/:id/reviews
Added route and controller for adding review to a product. If a user has already added a review or if the product can't be found an error will be returned. If the review is added successfully the rating, review count, and reviews array will all be updated in the database.