Pagination for Displayed Products
Added buttons that display how many pages of products there are. Clicking on the button navigates the user to '/page/:pageNumber'. When the query string hits the backend product controller, the Product model skips the listed products necessary and returns the products that should be displayed on that page number.